# Performance: use in-place matrix operations